+++ 3_etl_code/ETL_Orchestration/Scripts/setup_enviroments/setup_unit_test_enviroment.R
@@ -44,8 +44,20 @@ create_birth_mother_input_table(config)
 
 create_vision_input_table(config)
 
-# Step Six --------------------------------------------------------------
+# Step Five --------------------------------------------------------------
 # Create the necessary Kanta register table
 # Run function create_kanta_input_table.
 
 create_kanta_input_table(config)
+
+# Step Six --------------------------------------------------------------
+# Create the necessary drug events register table
+# Run function create_drug_events_input_table.
+
+create_drug_events_input_table(config)
+
+# Step Seven --------------------------------------------------------------
+# Create the necessary Spirometry register table
+# Run function create_spirometry_input_table.
+
+create_spirometry_input_table(config)