+++ NEWS.md
@@ -1,3 +1,11 @@
+# ETL dev
+
+# ETL ETLvDF14.1
+
+- Added HLA imputed data as non-standard concepts without any standard mapping and exist only in measurement table
+- Include only purchased and administered drugs; do NOT include prescription-only drugs.
+- Bug fixed: measurement unit column changed
+
 # ETL vDF13.3
 
 - ETL for FinnGen DF13 v3